completed project info popuup functionality and migrated delete project function to popup
