+++ kensa-ts/src/assets/styles/styles.scss
@@ -32,3 +32,23 @@ body {
 #delete-project {
     display: block;
 }
+
+#project-info {
+    &>div {
+        display: flex;
+        justify-content: space-between;
+        margin: 15px 0;
+    }
+
+    h4 {
+        margin: 10px 0;
+    }
+
+    span {
+        font-weight: normal;
+    }
+
+    button {
+        padding: 5px;
+    }
+}
